Remove FeatureBase.cpp
