+++ Gems/GenAIFramework/Code/genaiframework_private_files.cmake
@@ -15,8 +15,6 @@ set(FILES
     Source/Clients/GenAIFrameworkSystemComponentConfiguration.cpp
     Source/Clients/GenAIFrameworkSystemComponentConfiguration.h
 
-    Source/Feature/FeatureBase.cpp
-
     Source/O3DEAssistantFeature/O3DEAssistantFeature.cpp
     Source/O3DEAssistantFeature/O3DEAssistantFeature.h
 )
